Updates to filter

+++ src/main/java/com/cs334/project3/api/Controller.java
@@ -17,6 +17,7 @@ import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import java.time.ZonedDateTime;
 import java.util.List;
 
 @RestController
@@ -61,13 +62,26 @@ public class Controller {
     }
 
     @GetMapping(value="/groups", params= "userId")
-    public ResponseEntity<List<GroupDTO>> getGroupsForUser(@RequestParam Long userId){
-        try {
-            List<GroupDTO> gDTOList = groupMemberService.getGroupsWhereUserIsMember(userId);
-            return new ResponseEntity<>(gDTOList, HttpStatus.OK);
-        } catch (Exception e){
-            e.printStackTrace();
-            throw new InternalServerErrorException("Exception raised trying to retrieve group");
+    public ResponseEntity<List<GroupDTO>> getGroupsForUser(@RequestParam Long userId, HttpServletRequest request){
+        String token = request.getHeader("Authorization");
+        String jwt = null;
+        if(token.startsWith("Bearer ")){
+            jwt = token.substring(7);
+        } else {
+            throw new ResourceNotFoundException();
+        }
+        Long uid = getUserIdFromJWT(jwt);
+
+        if (uid.equals(userId)) {
+            try {
+                List<GroupDTO> gDTOList = groupMemberService.getGroupsWhereUserIsMember(userId);
+                return new ResponseEntity<>(gDTOList, HttpStatus.OK);
+            } catch (Exception e){
+                e.printStackTrace();
+                throw new InternalServerErrorException("Exception raised trying to retrieve group");
+            }
+        } else {
+            throw new ForbiddenException("User cannot access another users information");
         }
     }
 
@@ -108,17 +122,34 @@ public class Controller {
 
     ////////////////////Controller for groupmembers//////////////
     @PostMapping("/groupmember")
-    public ResponseEntity<GroupMembersDTO> addGroupMemberToGroupById(@RequestBody GroupRequestBody grbm){
-        UserDTO userDTO = userService.getUserById(grbm.getUserId());
-        if (userDTO == null){
-            throw new ResourceNotFoundException("No user with user ID "+grbm.getUserId()+" exists");
+    public ResponseEntity<GroupMembersDTO> addGroupMemberToGroupById(@RequestBody GroupRequestBody grbm, HttpServletRequest request) {
+        //Get userId from token:
+        String token = request.getHeader("Authorization");
+        String jwt = null;
+        if(token.startsWith("Bearer ")){
+            jwt = token.substring(7);
         } else {
-            try{
-                GroupMembersDTO gmdto = groupService.joinGroup(grbm.getGroupId(), grbm.getUserId(), grbm.isAdmin());
-                return new ResponseEntity<>(gmdto, HttpStatus.CREATED);
-            } catch(Exception e){
-                throw new InternalServerErrorException("User "+grbm.getUserId()+" could not be added to the group");
+            throw new ResourceNotFoundException();
+        }
+        Long uid = getUserIdFromJWT(jwt);
+
+        //Get userId from param:
+        Long userId = grbm.getUserId();
+
+        if (uid.equals(userId)) {
+            UserDTO userDTO = userService.getUserById(grbm.getUserId());
+            if (userDTO == null) {
+                throw new ResourceNotFoundException("No user with user ID " + grbm.getUserId() + " exists");
+            } else {
+                try {
+                    GroupMembersDTO gmdto = groupService.joinGroup(grbm.getGroupId(), grbm.getUserId(), grbm.isAdmin());
+                    return new ResponseEntity<>(gmdto, HttpStatus.CREATED);
+                } catch (Exception e) {
+                    throw new InternalServerErrorException("User " + grbm.getUserId() + " could not be added to the group");
+                }
             }
+        } else {
+            throw new ForbiddenException("User cannot access another users information");
         }
     }
 
@@ -135,14 +166,31 @@ public class Controller {
     }
 
     @PutMapping("/groupmember")
-    public ResponseEntity<GroupMembersDTO> updateGroupAdminByUser(@RequestBody GroupRequestBody grbm){
-        GroupMembersDTO gmDTO;
-        try{
-            gmDTO = groupMemberService.updateGroupAdminByUserId(grbm.getUserId(), grbm.getGroupId(), grbm.isAdmin());
-            return new ResponseEntity<>(gmDTO, HttpStatus.OK);
-        } catch(Exception e){
-            //TODO: Dom: Distinguish exception types
-            throw new InternalServerErrorException("Something went wrong, could not update admin");
+    public ResponseEntity<GroupMembersDTO> updateGroupAdminByUser(@RequestBody GroupRequestBody grbm, HttpServletRequest request){
+        //Get userId from token:
+        String token = request.getHeader("Authorization");
+        String jwt = null;
+        if(token.startsWith("Bearer ")){
+            jwt = token.substring(7);
+        } else {
+            throw new ResourceNotFoundException();
+        }
+        Long uid = getUserIdFromJWT(jwt);
+
+        //Get userId from param:
+        Long userId = grbm.getUserId();
+
+        if (uid.equals(userId)) {
+            GroupMembersDTO gmDTO;
+            try {
+                gmDTO = groupMemberService.updateGroupAdminByUserId(grbm.getUserId(), grbm.getGroupId(), grbm.isAdmin());
+                return new ResponseEntity<>(gmDTO, HttpStatus.OK);
+            } catch (Exception e) {
+                //TODO: Dom: Distinguish exception types
+                throw new InternalServerErrorException("Something went wrong, could not update admin");
+            }
+        } else {
+            throw new ForbiddenException("User cannot access another users information");
         }
     }
 
@@ -158,11 +206,7 @@ public class Controller {
         }
     }
 
-
     ////////////////////Controller for categories/////////////////////
-
-
-
     @GetMapping("/categories")
     @ResponseStatus(HttpStatus.OK)
     public ResponseEntity<List<CategoryDTO>> getAllCategories() {
@@ -175,48 +219,97 @@ public class Controller {
         }
     }
 
-
     ////////////////////Controller for friends//////////////
     @PostMapping("/friends")
     @ResponseStatus(HttpStatus.CREATED)
-    public ResponseEntity<FriendDTO> addingFriends(@RequestBody FriendRequestBody ids) {
-        FriendDTO friendDTO;
-        try {
-            friendDTO = friendService.addFriend(ids);
-            return new ResponseEntity<>(friendDTO, HttpStatus.CREATED);
-        } catch (Exception e) {
-            throw new InternalServerErrorException("Exception raised trying to add friend");
+    public ResponseEntity<FriendDTO> addingFriends(@RequestBody FriendRequestBody ids, HttpServletRequest request) {
+        //Get userId from token:
+        String token = request.getHeader("Authorization");
+        String jwt = null;
+        if(token.startsWith("Bearer ")){
+            jwt = token.substring(7);
+        } else {
+            throw new ResourceNotFoundException();
+        }
+        Long uid = getUserIdFromJWT(jwt);
+
+        //Get userId from param:
+        Long userId = ids.getUserId();
+
+        if (uid.equals(userId)) {
+            FriendDTO friendDTO;
+            try {
+                friendDTO = friendService.addFriend(ids);
+                return new ResponseEntity<>(friendDTO, HttpStatus.CREATED);
+            } catch (Exception e) {
+                throw new InternalServerErrorException("Exception raised trying to add friend");
+            }
+        } else {
+            throw new ForbiddenException("User cannot access another users information");
         }
     }
 
     @DeleteMapping("/friends")
     @ResponseStatus(HttpStatus.OK)
-    public void removeFriend(@RequestBody FriendRequestBody ids) {
-        try {
-            Friend friend = friendService.getFriendById(ids.getUserId(), ids.getFriendId());
-            friendService.deleteFriend(friend);
-        } catch (Exception e) {
-            throw new InternalServerErrorException("Exception raised trying to delete friend");
+    public void removeFriend(@RequestBody FriendRequestBody ids, HttpServletRequest request) {
+        //Get userId from token:
+        String token = request.getHeader("Authorization");
+        String jwt = null;
+        if(token.startsWith("Bearer ")){
+            jwt = token.substring(7);
+        } else {
+            throw new ResourceNotFoundException();
+        }
+        Long uid = getUserIdFromJWT(jwt);
+
+        //Get userId from param:
+        Long userId = ids.getUserId();
+
+        if (uid.equals(userId)) {
+            try {
+                Friend friend = friendService.getFriendById(ids.getUserId(), ids.getFriendId());
+                friendService.deleteFriend(friend);
+            } catch (Exception e) {
+                throw new InternalServerErrorException("Exception raised trying to delete friend");
+            }
+        } else {
+            throw new ForbiddenException("User cannot access another users information");
         }
     }
 
     @GetMapping(value="/friends", params="userId")
-    public ResponseEntity<List<FriendDTO>> getFriendsOfUser(@RequestParam Long userId) {
-        List<FriendDTO> friendDTOList;
-        try {
-            friendDTOList = friendService.getFriendsByUserId(userId);
-            return new ResponseEntity<>(friendDTOList, HttpStatus.OK);
-        } catch (Exception e) {
-            throw new InternalServerErrorException("Friends for " + userId + " could not be retrieved");
+    public ResponseEntity<List<FriendDTO>> getFriendsOfUser(@RequestParam Long userId, HttpServletRequest request) {
+        //Get userId from token:
+        String token = request.getHeader("Authorization");
+        String jwt = null;
+        if(token.startsWith("Bearer ")){
+            jwt = token.substring(7);
+        } else {
+            throw new ResourceNotFoundException();
+        }
+        Long uid = getUserIdFromJWT(jwt);
+
+        if (uid.equals(userId)) {
+            List<FriendDTO> friendDTOList;
+            try {
+                friendDTOList = friendService.getFriendsByUserId(userId);
+                return new ResponseEntity<>(friendDTOList, HttpStatus.OK);
+            } catch (Exception e) {
+                throw new InternalServerErrorException("Friends for " + userId + " could not be retrieved");
+            }
+        } else {
+            throw new ForbiddenException("User cannot access another users information");
         }
     }
 
     ////////////////////Controller for posts/////////////////////
-
-    @GetMapping("/posts")
-    public ResponseEntity filter(@RequestBody FilterPostsRequestBody criteria) {
+    @GetMapping(value="/posts", params= {"userId", "filterUsedId", "groupId", "after", "time", "radiusKm", "longitude", "latitude"})
+    public ResponseEntity filter(@RequestParam Long userdId, @RequestParam Long filterUsedId,
+                                 @RequestParam Long groupId, @RequestParam Boolean after,
+                                 @RequestParam ZonedDateTime time, @RequestParam Double radiusKm,
+                                 @RequestParam Double longitude, @RequestParam Double latitude) {
         try {
-            List<PostDTO> posts = postService.filterPosts(criteria);
+            List<PostDTO> posts = postService.filterPosts(userdId, filterUsedId, groupId, after, time, radiusKm, longitude, latitude);
             return new ResponseEntity<>(posts, HttpStatus.OK);
         } catch (Exception e) {
             e.printStackTrace();
@@ -225,13 +318,30 @@ public class Controller {
     }
 
     @PostMapping("/posts")
-    public ResponseEntity<PostDTO> addPost(@RequestBody PostRequestBody ids) {
-        try {
-            PostDTO pdto = postService.createPost(ids);
-            return new ResponseEntity<>(pdto, HttpStatus.CREATED);
-        } catch (Exception e) {
-            e.printStackTrace();
-            throw new InternalServerErrorException("Error in creating post: " + e.getMessage());
+    public ResponseEntity<PostDTO> addPost(@RequestBody PostRequestBody ids, HttpServletRequest request) {
+        //Get userId from token:
+        String token = request.getHeader("Authorization");
+        String jwt = null;
+        if(token.startsWith("Bearer ")){
+            jwt = token.substring(7);
+        } else {
+            throw new ResourceNotFoundException();
+        }
+        Long uid = getUserIdFromJWT(jwt);
+
+        //Get userId for param:
+        Long userId = ids.getUserId();
+
+        if (uid.equals(userId)) {
+            try {
+                PostDTO pdto = postService.createPost(ids);
+                return new ResponseEntity<>(pdto, HttpStatus.CREATED);
+            } catch (Exception e) {
+                e.printStackTrace();
+                throw new InternalServerErrorException("Error in creating post: " + e.getMessage());
+            }
+        } else {
+            throw new ForbiddenException("User cannot access another users information");
         }
     }
 
@@ -248,13 +358,27 @@ public class Controller {
 
     ////////////////////Controller for users/////////////////////
     @GetMapping(value="/users", params="userId")
-    public ResponseEntity<UserDTO> getUserById(@RequestParam Long userId){
-        UserDTO userDTO;
-        try{
-            userDTO = userService.getUserById(userId);
-            return new ResponseEntity<>(userDTO, HttpStatus.OK);
-        } catch(Exception e){
-            throw new ResourceNotFoundException("No user with user ID "+userId+" exists");
+    public ResponseEntity<UserDTO> getUserById(@RequestParam Long userId, HttpServletRequest request){
+        //Get userId from token:
+        String token = request.getHeader("Authorization");
+        String jwt = null;
+        if(token.startsWith("Bearer ")){
+            jwt = token.substring(7);
+        } else {
+            throw new ResourceNotFoundException();
+        }
+        Long uid = getUserIdFromJWT(jwt);
+
+        if (uid.equals(userId)) {
+            UserDTO userDTO;
+            try {
+                userDTO = userService.getUserById(userId);
+                return new ResponseEntity<>(userDTO, HttpStatus.OK);
+            } catch (Exception e) {
+                throw new ResourceNotFoundException("No user with user ID " + userId + " exists");
+            }
+        } else {
+            throw new ForbiddenException("User cannot access another users information");
         }
     }
 
@@ -269,12 +393,26 @@ public class Controller {
     }
 
     @DeleteMapping(value="/users", params="userId")
-    public ResponseEntity<UserDTO> deleteUserById(@RequestParam Long userId){
-        try{
-            UserDTO userDTO = userService.deleteUserById(userId);
-            return new ResponseEntity<>(userDTO,HttpStatus.OK);
-        } catch(Exception e){
-            throw new ResourceNotFoundException("No user with ID "+userId+" exists");
+    public ResponseEntity<UserDTO> deleteUserById(@RequestParam Long userId, HttpServletRequest request){
+        //Get userId from token:
+        String token = request.getHeader("Authorization");
+        String jwt = null;
+        if(token.startsWith("Bearer ")){
+            jwt = token.substring(7);
+        } else {
+            throw new ResourceNotFoundException();
+        }
+        Long uid = getUserIdFromJWT(jwt);
+
+        if (uid.equals(userId)) {
+            try {
+                UserDTO userDTO = userService.deleteUserById(userId);
+                return new ResponseEntity<>(userDTO, HttpStatus.OK);
+            } catch (Exception e) {
+                throw new ResourceNotFoundException("No user with ID " + userId + " exists");
+            }
+        } else {
+            throw new ForbiddenException("User cannot access another users information");
         }
     }
 
@@ -292,17 +430,33 @@ public class Controller {
     }
 
     @PutMapping(value="/users")
-    public ResponseEntity<UserDTO> updateUser(@RequestBody UpdateDispnameRequestBody udrb){
-        try{
-            UserDTO userDTO = userService.updateUser(udrb);
-            return new ResponseEntity<>(userDTO, HttpStatus.OK);
-        } catch(Exception e){
-            throw new ResourceNotFoundException("No such user exists");
+    public ResponseEntity<UserDTO> updateUser(@RequestBody UpdateDispnameRequestBody udrb, HttpServletRequest request){
+        //Get userId from token:
+        String token = request.getHeader("Authorization");
+        String jwt = null;
+        if(token.startsWith("Bearer ")){
+            jwt = token.substring(7);
+        } else {
+            throw new ResourceNotFoundException();
+        }
+        Long uid = getUserIdFromJWT(jwt);
+
+        //Get userId for param:
+        Long userId = udrb.getUserid();
+
+        if (uid.equals(userId)) {
+            try {
+                UserDTO userDTO = userService.updateUser(udrb);
+                return new ResponseEntity<>(userDTO, HttpStatus.OK);
+            } catch (Exception e) {
+                throw new ResourceNotFoundException("No such user exists");
+            }
+        } else {
+            throw new ForbiddenException("User cannot access another users information");
         }
     }
 
     ////////////////////Register, logout/////////////////////
-
     @PostMapping(value="/register")
     public ResponseEntity<CreateUserStatus> registerNewUser(@RequestBody CreateUserRequestBody params){
         try{
